Zipping can now be performed with target path specified with content:// uri.

+++ android/src/main/java/com/rnziparchive/ZipTask.java
@@ -1,12 +1,18 @@
 package com.rnziparchive;
 
+import android.net.Uri;
+import android.support.v4.provider.DocumentFile;
+
 import com.facebook.react.bridge.Promise;
+import com.facebook.react.bridge.ReactApplicationContext;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
+import java.io.OutputStream;
+import java.net.URLConnection;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.zip.ZipEntry;
@@ -17,19 +23,21 @@ public class ZipTask {
   private final String[] files;
   private final String fromDirectory;
   private final Promise promise;
+  private final ReactApplicationContext mCtx;
   private static final int BUFFER_SIZE = 4096;
 
   private long bytesRead = 0;
   private long totalSize;
-  private RNZipArchiveModule cb;
+  private RNZipArchiveModule rnZipArchiveModule;
   private String threadError;
 
-  public ZipTask(String[] files, String destFile, String fromDirectory, Promise promise, RNZipArchiveModule cb) {
+  public ZipTask(String[] files, String destFile, String fromDirectory, Promise promise, RNZipArchiveModule rnZipArchiveModule, ReactApplicationContext mCtx) {
     this.destFile = destFile;
     this.files = files;
     this.fromDirectory = fromDirectory.endsWith("/") ? fromDirectory : fromDirectory + "/";
     this.promise = promise;
-    this.cb = cb;
+    this.rnZipArchiveModule = rnZipArchiveModule;
+    this.mCtx = mCtx;
   }
 
   public void zip() {
@@ -41,30 +49,45 @@ public class ZipTask {
 
     Thread t = new Thread(new Runnable() {
       public void run() {
+        Uri destUri = Uri.parse(destFile);;
+        OutputStream dest;
+        BufferedInputStream origin;
+        boolean fromContentUri = "content".equals(destUri.getScheme());
+        final String progressKey = fromContentUri ? fromDirectory : destFile;
         try {
-          if (destFile.contains("/")) {
-            File destDir = new File(destFile.substring(0, destFile.lastIndexOf("/")));
-            if (!destDir.exists()) {
-              //noinspection ResultOfMethodCallIgnored
-              destDir.mkdirs();
-            }
-          }
-
-          if (new File(destFile).exists()) {
-            //noinspection ResultOfMethodCallIgnored
-            new File(destFile).delete();
-          }
-
           final long totalUncompressedBytes = getUncompressedSize(files);
 
-          BufferedInputStream origin;
-          FileOutputStream dest = new FileOutputStream(destFile);
+          if (fromContentUri) {
+            Uri dirUri = Uri.parse(destFile);
+            Uri destPath = Uri.parse(fromDirectory);
+
+            DocumentFile pickedDir = DocumentFile.fromTreeUri(mCtx, dirUri);
+
+            String mimeType = URLConnection.guessContentTypeFromName(destPath.getLastPathSegment());
+            Uri targetFileUri = pickedDir.createFile(mimeType, destPath.getLastPathSegment()).getUri();
+            dest = mCtx.getContentResolver().openOutputStream(targetFileUri);
+          } else {
+            if (destFile.contains("/")) {
+              File destDir = new File(destFile.substring(0, destFile.lastIndexOf("/")));
+              if (!destDir.exists()) {
+                //noinspection ResultOfMethodCallIgnored
+                destDir.mkdirs();
+              }
+            }
+
+            if (new File(destFile).exists()) {
+              //noinspection ResultOfMethodCallIgnored
+              new File(destFile).delete();
+            }
+
+            dest = new FileOutputStream(destFile);
+          }
 
           ZipOutputStream out = new ZipOutputStream(new BufferedOutputStream(dest));
 
           byte data[] = new byte[BUFFER_SIZE];
 
-          cb.updateProgress(0, 1, destFile); // force 0%
+          rnZipArchiveModule.updateProgress(0, 1, progressKey); // force 0%
           for (int i = 0; i < files.length; i++) {
             String absoluteFilepath = files[i];
 
@@ -80,7 +103,7 @@ public class ZipTask {
               timer.scheduleAtFixedRate(new TimerTask() {
                 @Override
                 public void run() {
-                  cb.updateProgress(bytesRead, totalUncompressedBytes, destFile);
+                  rnZipArchiveModule.updateProgress(bytesRead, totalUncompressedBytes, progressKey);
                 }
               }, 200, 200);
 
@@ -92,14 +115,14 @@ public class ZipTask {
               origin.close();
             }
           }
-          cb.updateProgress(1, 1, destFile); // force 100%
+          rnZipArchiveModule.updateProgress(1, 1, progressKey); // force 100%
           out.close();
         } catch (Exception ex) {
           ex.printStackTrace();
-          cb.updateProgress(0, 1, destFile); // force 0%
-          promise.reject(null, String.format("Couldn't zip %s", destFile));
+          rnZipArchiveModule.updateProgress(0, 1, progressKey); // force 0%
+          promise.reject(null, String.format("Couldn't zip %s", progressKey));
         }
-        promise.resolve(destFile);
+        promise.resolve(progressKey);
       }
     });
 
@@ -107,7 +130,7 @@ public class ZipTask {
     t.start();
   }
 
-    /**
+  /**
    * Return the uncompressed size of the ZipFile (only works for files on disk, not in assets)
    *
    * @return -1 on failure
